Add accessors, topper/bottomer customization
In the default configuration, sorting is manipulated with the traditional 'first' and 'last' characters of SPACE and TILDE.  In circumstances where these characters will not work, you can now specify different characters for both the 'topper' and the 'bottomer'.  These changes can be made individually at the instance level or globally at the class level.

This commit also adds the ability to change the call number text for a given instance, and tests for all the new features.



git-svn-id: https://library-callnumber-lc.googlecode.com/svn/trunk@34 cf210cc4-13c3-11de-bcda-2bb9902f8029