Issue 7 - Removing forced normalization
There are times[1] when calling code may not want forced changes to the callnumber
it passes off to the Library::Callnumber::LC object. Since we already handle
the conversion to uppercase in the normalize method, there is really no reason
to force uppercase on the creation of an L::C::LC object. This patch removes
the code which forces uppercase.

This patch also moves the forced addition of a decimal to the first cutter. Again,
the caller should not be handed back a normalized callnumber unless specifically
asked for.

It also adds another number to t/t_Labels_split_lccn.t

[1] http://bugs.koha-community.org/bugzilla3/show_bug.cgi?id=10821